resource: automatically discharge all attached event sources when destroying resource
This change adds more automatism in destroying all attached event sources
when we destroy the resource so the caller/code outside doesn't need to
bother about this. All event sources attached to the resource are
automatically enlisted inside the resource when they are created and then
when the resource is being destroyed, all the event sources are destroyed
as well.

However, it is still possible to destroy individual event sources outside
of the resource itself if we need that - for example, if we no longer use
the event source but the resource is still used and active.