resource: command: fix invalid read of connection buffer
After we receive data to connection buffer in connection resource
about a command to execute together with all its arguments, we instantiate
a command resource to actually process the command. However, we can
access the connection buffer only during command resource initialization
because after this point, the connection buffer is reset.

We already parsed the commmand arguments (the environment key-value
pairs) during command resource initialization, but we didn't copy the
command header - we directly used the connection buffer which lead to
invalid memory access when we tried to read the header while processing
the command later after the command resource initialization (e.g. the
_cmd_handler function where we're checking the command request header).