tests: add unit test directory and buffer test
This adds a tests directory, for unit tests, which can be run with
"make check".  It also adds a currently failing test to check if
buffer_fmt_add works correctly on linear buffers that already have
enough size to hold the data.

This testing setup is pretty sub-optimal. Cmocka doesn't report the
individual test results in the console output. Instead, you need to look
the test log file.  It is able to output correctly with the autotools
testing setup, but that takes some additional work.