iface: fix _do_service_link_notify() buffer creation
If _do_service_link_notify() is called with a NULL fmt for a
notification that wants and argument string, it will crash. Also, if
_do_service_link_notify() is called with multiple notifications (except
for SERVICE_NOTIFICATION_STATUS and SERVICE_NOTIFICATION_ERRNO) it will
have nulls in the middle of the string it passes to sd_notify(). Fix
these issues.