misc: buffer: rename buffer_reset to buffer_reset_init; buffer_reset without params now
Usually, we don't want to change params on buffer reset and we reuse the ones we already
supplied during buffer_create. This patch moves the original buffer_reset functionality
to new buffer_reset_init - this one takes the initial_size and alloc_step params.

The buffer_reset doesn't have the init params, just the buffer itself which
now remembers all the params it needs to do the reset. For this, we added the
'initial_size' inside buffer structure so we can use it on reset.

As a side effect, with the initial_size remembered inside the buffer
structure, we can now easily see how many times we have extendend the buffer
from its initial size right from the buffer structure itself. We can use
this in the future to gather statistics and optimize buffers.