modules: ubridge-cmd: multipath: store more key values
When the multipath_component module finds a valid or possibly valid path
device, it now stores the WWID, which will be used in a future patch to
help determine if a possibly valid device should be marked as valid
path. It also stores the result of checking if the device is a valid
path to the device namespace, so it can be looked up later by other
devices.