resource: worker-control: fix unintentional use of worker wire spec extensions
Add 'used' switch under struct worker_wire_spec and its 'ext' field to
avoid unintentional use of the extensions if they're not initialized
implicitly.

For example, if we have worker_wire_spec.ext.pipe.fd_redir = 0 (which is
the case when we initialize the whole struct with zeroes) and not -1 as
it should, the code parsing the extension part then incorrectly tries to
redirect FD number 0 ('0' is proper FD number, -1 would normally direct
the code to ignore it).

With the worker_wire_spec.ext.used, we need to explicitly set that for
the 'ext' part of the worker_wire_spec to be taken into account and so
it makes it more visible that we need to initialize the struct
worker_wire_spec more thoroughly.