resource: worker-control: set lower priority for _on_worker_proxy_child_event handler
Currently, the _on_worker_proxy_child_event handler is responsible for
destroying the worker_proxy once we get notified that associated worker
has exited.

However, if we still have remaining events to be processed, like
receving data from channels, we need to be able to finish these before
we destroy the worker_proxy (which would also destroy the channels). For
this reason, we need to run the _on_worker_proxy_child_event with lower
priority than other event handlers.

Without this patch, the problem manifests when we have interleaved
events. A good example is when we have WORKER_TYPE_EXTERNAL worker
where we redirect output to worker proxy through a channel and the
channel needs to extend the buffer in chunks to collect complete output.
Data reception for each chunk is then a separate event. And these may
interleave with _on_worker_proxy_child_event which would cause
destruction of the worker_proxy before we had a chance to receive all
chunks from the channel.