tools: rpm: configure: install 'usid' to udev directory
'usid' is supposed to be called from within a udev rule only, so place it to
directory where udev helpers are (usually /usr/lib/udev/).

Now that the 'usid' tool is in udev directory, we can call it without
specifying full path in the udev rule and that also means we don't need
to do any path substitutions here and that simplifies the makefile too.