resource: ubridge: make 'init' and 'exit' internal phase a part of the _cmd_exec_scan loop
The 'init' phase containing initialization and 'exit' the finalization
of the whole 'scan' command is like any other phase of this command,
it's just internal, not exposed as module hooks.

One of the advantages is that we can apply flags to check allowed
operations against like we do for other phases already (the cmd_reg.flags).
Another advantage is that the code is more consistent and the same logic
applies for all phases, including the internal ones.