resource: use temporary reference to a resource in sid_resource_create
When we are in the process of creating a resource, we call out
type->init callback for any type-specific initialization. That init can
take references to the resource (e.g. as part of creating an iterator
inside). If there is any failure during the initialization and if we drop
the references taken down to 0, the automatic resource destruction would
trigger. We would like to avoid this and do the actual resource
destruction in the outermost sid_resource_create function for clarity
and to avoid confusion.