resource: module-registry: support submodules; add module_registry_add_module_subregistry
An instance of a module-registry can now be attached to a module, hence
creating a submodule registry for given module.

After attaching the submodule registry, we have:

  - the submodule registry's resource is attached as child to module's resource

  - the submodule registry's base name is prefixed with module's full name
    and separated by '/' character, for example, after adding new module
    registry named "sub" under module "type/some_mod" we end up with
    "type/some_mod/sub" as a name for the new module registry

  - any module registered under the "type/some_mod/sub" registry will
    now inherit this base for its full name, for example
    "type/some_mod/sub/some_sub_mod"

This way, we practically store the whole path within a module tree in
module's full name and making it possible to search through the module
tree just by providing module's full name.