util: add util_mem_t *mem param to all functions using allocated memory
All functions that need to use allocated memory and they provide a possibility
to use preallocated memory contain 'util_mem_t *mem' parameter to pass this
preallocated memory for use. The 'util_mem_t' is a simple structure to pass
memory base and size.

If 'mem' or 'mem->base' is NULL, the functions allocate the memory by
themselves.