modules: ucmd: block: blkid: fix detected module name to not include '.so' suffix
The blkid module stores detected module name for next layer in
SID_NEXT_MOD (SID_UCMD_KEY_DEVICE_NEXT_MOD) KV record that the core
reads and loads 'type' module based on this name to handle the next
layer.

This is a regression and it was introduced by commit
070e8cd7b4bda4cc1e2051627ed43da80a60927a since which module names do not
include common prefixes and suffixes and when looking for a module by
its name, we need to provide the name without the prefix and/or suffix.