resource: ubridge: send reply when creating command resource fails
When a worker fails trying to create a command resource, it never
notifies the client. This means that the client keeps waiting for a
response from the worker, and the worker keeps waiting for the client to
close the socket. Instead, the worker should send the client a failure
reply.