resource: ubridge: clean up _cmd_handler error paths
If the sid fails to execute the worker command, it should set
COMMAND_STATUS_FAILURE, and shouldn't export the kv_store, since it
may have had an error updating its local copy. If it doesn't
understand the protocol the client is communicating with, it should
simply abort the connection, since it doesn't know now to communicate
with the client.