resource: worker-control: also copy channel spec ids from params to worker control resource
The worker_control resource initialization code already copied channel specs
from given resource params to its own field in 'struct worker_control' but
it did not copy the IDs which are part of the specs (it only took the references).

We need to do complete deep copy here and so including the copy of spec IDs
because worker_control resource initializaton params could be inaccessible after
the worker_control is initialized.