resource: worker-control: hand over channel specs from worker_control to worker after forking
When we create a new worker, there's 'worker_proxy' internal resource
on parent side that is used for monitoring the worker and it also contains
communication channels to/from the worker. The 'worker_proxy' resource is
always attached to 'worker_control'.

The 'worker_control' contains channel specs that 'worker_proxy' references
so we don't need to worry about references from one resource to the other
here.

On worker side, there's 'worker' resource but without any 'worker_control'
which is not used in worker. That means we need to hand over the channel
specs from 'worker_control' to 'worker' because the 'worker_control' is
destroyed right after the worker is set up. Otherwise, any references from
'worker' to 'worker_control' would become invalid.

This patch this handover which was missing.