resource: worker-control: use sid_resource_unref instead of sid_resource_destroy in worker_control_get_new_worker
When creating top-level 'sid' resource in daemon code, we take a direct
reference to it because there's no other resource using it above. When we
want to destroy the resource in worker_control_get_new_worker after forking
in the worker part of the code, we need to call sid_resource_unref so the
direct reference is dropped.