resource: use signalfd() to map signal handlers
Create a signal file descriptor to handle the signals via signalfd().
Add the file descriptor to a systemd IO event source via sd_event_add_io().

The advantage of using signalfd() is when the process forks - the child
can cleanup/close the file descriptor without impacting the signal
mappings of the parent.  It also allows the child the remap the signals
the appropriate signal handler.

Signed-off-by: Todd Gill <tgill@redhat.com>