sidctl: add stats command
The "sidctl stats" command prints various memory usage stats from the
key-value store.  It shows total size of the keys, the hash metadata
(including the keys), the internal and external values, and the number
of key value pairs. The value stats show both the total size of the
memory used to hold the values, including metadata, and the size of just
the value data itself.