resource: remove sid_resource_dump_all_in_dot
The sid_resource_dump_all_in_dot was only temporary for debug purposes.

We have 'sidctl tree' now that can cover this functionality. The 'sidctl tree'
output can be converted to 'dot' graph if we want to (we can add a
script to do the translation).

(I'm keeping the sid_resource_dump_all_in_dot code in place for now,
just commented out and removed from resource.h interface. When we have
the translator in place, we can remove this part completely.)