resource: ubridge: mmap correct size to sync kv_store
When mmaping the memfd that holds the exported kv_store information, SID
was only mmaping the size of the kv_store data, without including the
this size_t which stored the overall size.  However the mmap started at
the beginning of the file (as it needs to, since the mmap offset must be
in increments of the page size). This means that SID was requesting too
little memory in the mmap. Fortunately, mmap returns memory in page
sized chunks, so unless the the kv_store data took up exactly a page,
mmap would return enough extra to allow all the exported kv_store data
to be accessed. But there's no reason why the kv_store data size can't
end up being a multiple of the page size, so SID needs to request the
correct size.