buffer: add 'get_fd' buffer_type fn and update size prefix for BUFFER_MODE_SIZE_PREFIX
When calling buffer_get_fd, it is expected the FD is going to be send
over to a consumer. The consumer surely needs to see up-to-date value in
size prefix if we have a buffer with BUFFER_MODE_SIZE_PREFIX mode. The
size prefix is handled by each buffer type in its own way, so we need a
type-specific hook for the type to do any updates needed before the FD
is provided to the caller - that's the purpose of the new 'get_fd'
buffer type hook.