buffer: allow adding NULL data to linear buffers
With vector buffers, it is possible to first allocate space and add it
to the buffer, and then fill it. This wasn't possible with linear
buffers, which required already existing space to copy the data from.
buffer_add() should allow a NULL data pointer, and simply memset the
allocated space to zero in this case. This allows linear buffers to be
used with functions that are designed to fill preallocated memory.