resource: ubridge: use single place to write result and export buffer
Add 'exp_buf' (export buffer) to struct sid_ucmd_ctx and use single
place to write the result buffer as well as export buffer - the
_cmd_handler fn.

The 'res_buf' (result buffer) is the usual result which is directly
sent over to client and 'exp_buf' (export buffer) is meant to be sent
by means of sending the backing FD over a unix socket.

Add 'CMD_KV_EXPORT_CLIENT' cmd_reg flag to make a difference between
commands that need to send records to an external client (like 'sidctl dump')
and the ones that need to send them internally to the main SID process
(like 'usid scan').