iface: usid: simplify usid_req interface
usid_req() has some issues as a public library interface. It prints
output messages which the caller may not want. I requires callers to use
the buffer interface to pass in additional data and read the results,
which means that callers need to pull in a bunch of functions unrelated
to SID. Finally, callers are expected to deal with the fact that
sometimes the data is returned in the resp_buf, and sometimes it is in
returned in the resp_fd, and free these resources correctly when they're
done.

With these changes, usid_req() no longer prints any output, or requires
callers to care about where the data is stored. They can simply access
it and free it with the usid_result_* functions. The code still uses
buffers under the surface, so callers still need to link against the
other SID libraries. However this requirement can be removed by later
changes.