tools: sidctl: remove unneeded buffer copying
instead of writing all the SID daemon output to a buffer, and then
writing the buffer to stdout, sidctl can simply print the daemon
response directly.  However, the daemon wasn't previously NULL
terminating its response. This was correct when sidctl was writing the
output directly to stdout, but now sidctl is treating the output as
a string, which needs NULL termination.