resource: ubridge: add separate dev_num besides dev_id in struct sid_ucmd_ctx
This is a starting point for SID's persistent device IDs. We are aiming for a persistent
device ID which could be rematched if we persist the SID KV store with device records.
Then, even though major:minor could have changed, as well as other names/aliases for a
device while SID was not listening for events, the device ID should remain same.

The actual ID rematching logic will be a part of future patches.