iface: rename commands dump -> dbdump, stats -> dbstats, tree -> resources
Make it clear that 'dump' and 'stats' commands are related only to database
matters only by prefixing them with 'db' so we have 'dbdump' and 'dbstats' now.

Rename 'tree' command to 'resources' as its purpose is to show current
resources. The 'tree' might be confused by something like "device tree"
that we may potentially use later.