worker-control: use SOCK_STREAM for comms between main and worker process
For SOCK_DGRAM, we'd need to know the exact message boundary for us to read
it in one go. This can't work with the loop where we first read the message
size and then the data itself in another iteration - the data is already
discarded. Therefore, use SOCK_STREAM here.

Maybe SOCK_SEQPACKET would be more suitable here, but that would require
some enhancements in buffers too so we could use sendmsg/recvmsg instead
of pure write/read.

Also, use AF_UNIX instead of AF_LOCAL (which is a synonym anyway) just
for consistency with the rest of the code.