resource: ubridge: check client permission for running a command before creating the command resource
We already know the command type before creating the command resource so
check client's permission for running a command before creating the
command resource itself.

This also avoids referencing connection FD in _init_command and relaxes
dependency of command resource on connection resource (though there is
still one more such dependency in _cmd_handler when writing response
back to the connected client - we'll try to cover that in a patch later).