resource: ubridge: add KV_SYNC flag for KV_SYNC and KV_PERISTENT difference
Add new KV_SYNC kv store value flag so we can make a difference between
a "request for syncing the snapshot value with main kv store" and
"making the record persistent". When marking the record as persistent,
it will be counted in for dumping to a persistent store and hence able
to be imported as part of next database generation.

This patch also reflags current KV_PERSISTENT flag throughout code to
KV_SYNC which is what it is meant to be now.