base: comms: fix setting unix socket pathname
memccpy() copies a buffer either through a given character or till the
given length, whatever comes first.  So this code was treating path_len
as a character to check for, to end the memcpy. Instead, verify that
path_len is smaller than sun_path, and then copy than many bytes.