resource: prepare for changing how workers startup
In preparation for future patches, change the _get_worker() and
worker_control_get_new_worker() functions to return both a return code
and a resource. Right now, when the functions return success, the
returned resource will always be set to the worker_proxy resource, but
future patches will change this.