resource: worker-control: cleanup before starting worker
The worker process was running its event loop while in an event handler
for the main process event loop. This made it impossible to free the
main process resources that it inherited when it forked.

To avoid this defer actually creating the worker resource and cleaning
up the inherited resources until after exitting the old event loop.