resource: worker-control: make all fd redirects effect the worker
WORKER_WIRE_PIPE_TO_WORKER channels were redirecting file descriptors on
the proxy process, instead of the worker process. It seems more useful
to have an external worker process able to read from the proxy on
STDIN_FILENO, than for the proxy process to pick which fd is uses to
write to a random file descriptor of the external worker process.