resource: worker-control: fix segfault on dropped channel and NULL input buffer
The _on_worker_proxy_channel_event and _on_worker_channel_event event
loop callbacks are also called if there's an error detected on the
channel, not only when data is available for reception.

Also, these callbacks are used for channels where we do not expect
incoming data (which we collect in chan->in_buf) and so the input
buffer for such channels is not set at all (chan->in_buf == NULL).

For example, this is the case with the proxy side (write end) of the
WORKER_CHANNEL_PIPE_TO_WORKER. The worker could be brought down/crashed,
breaking the pipe which would emit the EPOLLERR on the proxy side and
then it's caught in _chan_buf_recv.

If that's the case, we have to avoid trying to reset the NULL
chan->in_buf.