resource: ubridge: fix dbdump command to really dump all records; add CMD_KV_EXPORT_ALL flag
We used KV_EXPORT_SYNC (matching with KV_SYNC KV record flag) and
CMD_KV_EXPORT_PERSISTENT (matching with KV_PERSISTENT KV record flag) to check
for KV records which are meant to be included in export buffers.

However, for commands like "dump", we want all KV records to be exported,
not just the ones marked with SYNC and/or PERSISTENT flag - there are records
which do not have any of the two set and still we want them to be included in
the dump.

For this purpose, we have now a new flag "CMD_KV_EXPORT_ALL" to cover
all records.