resource: ubridge: fix buffer usage to count with possible memory allocation movement (part 1)
Buffer's underlying 'realloc' as well as 'mremap' (with MREMAP_MAYMOVE)
will end up moving existing memory content if there's no contiguous space
for the extension of the existing memory block.

This patch uses sid_buffer_count to get current head offset in buffer's
memory so we can use sid_buffer_rewind instead of sid_buffer_rewind_mem
at the end of functions which use a buffer for temporary allocations.

Also, if we're returning a pointer to memory with the beginning of data
that was collected with sid_buffer_add throughout a function execution,
do remember initial offset instead of memory start pointer. Then at the
end of the function, get buffer's data and add the initial offset to it
to get the actual pointer to return (currently this was the case of
_buffer_get_vvalue_str).

This way, the pointers we're working with are correct even after
realloc/mremap moving memory.

Other (more complex) cases will be patched in subsequent patches.