base: buffer: return mem pointer and pos within buffer on demand for sid_buffer_{add,fmt_add,vfmt_add}
Change sid_buffer_add, sid_buffer_fmt_add and sid_buffer_vfmt_add so
that they return pointer to the start of the buffer memory which is
used for current addition only on demand - add 'mem' output arg for
this.

Also, make it possible to return absolute position within buffer memory
(not including any metadata like size prefix) - add 'pos' output arg
for this. This way, it's easy to remember the position in the caller
so it can use it for either:

  - sid_buffer_rewind
  - sid_buffer_get_data + pos for the offset within data to get an
    individual item

The 'ret_code' which was output arg before is now function's return code
directly.

This change will also make it easier to track whether we're only
accumulating to buffer memory or whether we're also using the value
outside (in which case, we request the memory pointer to be returned).

The memory usage tracking will then make it possible to detect situations
where we're using the buffer memory while there's a risk of the
realloc/mremap to get called later when there are more subsequent
sid_buffer_{add,fmt_add,vfmt_add} calls. The realloc/mremap could cause
any existing pointers to buffer memory to be invalidaded if the memory
segment needs to be moved due to the extension (...if the extension
causes the memory segment to not fit anymore so that it still remains
as one linear uninterrupted memory segment).