base: buffer: fix checking and setting buffer memory usage mark in sid_buffer_{add,fmt_add,vfmt_add}
If we called sid_buffer_{add,fmt_add,vfmt_add} with possible internal
realloc/mremap, we'd invalidate any PREVIOUS memory pointers returned
and hence we can't conditionalize the check in these functions based on
whether we're NOW interested in returning the mem pointer or not. We
just check for the mark existence.

On the other hand, if we're not interested in returning the mem pointer,
we shouldn't be setting the mark.