internal: formatter: unbind buffer memory pointer in _print_binary after use
The _print_binary calss sid_buffer_add just to reserve a memory segment in
buffer and get the pointer to the memory. Then, it prints the encoded
binary value to the buffer memory as pointed by the pointer and then
the pointer is not used anymore, hence it can be marked as unbound.