resource: ubridge: check if index is needed inside KV store callbacks
The KV store callbacks have access to both old and new records and so
they can easily check whether we also need to create or remove an index
for given key based on record content.

The result is then returned in struct kv_update_arg.ret_code that we can
check after the kv-store function with the callback is finished.