internal: bptree: generalize bptree_update to also handle removals
bptree_update now accepts removals. This way, it's possible to call
bptree_update_fn_t type callback before the actual removal so we can
do further cleanup of embedded structures inside a record, if needed,
or confirm the removal.

The bptree_update_fn_t type callback now also provides reference counter
for the existing record which is mostly important if we want to do a
cleanup of embedded structures inside a record only when last record
reference is just about to be dropped.