internal: bptree: add bptree_destroy_with_fn
Just like bptree_destroy, but allows for calling a callback function
before each record is actually destroyed. This is mostly useful when the
tree contains key aliases and we want to do a cleanup for embedded
structures inside a record and we need to do that only when very last
record reference is dropped.