resource: ubridge: add 'index' arg to _kv_delta_set
With KV store snapshot, we need to create an index for all records to be synced
with main KV store for quick lookup when syncing. With main KV store, we don't
create such an index. The _kv_delta_set is used in both the worker (with KV
store snapshot) and main process (with main KV store) - add 'index' bool arg
to _kv_delta_set to make a difference.