internal: bptree: add keys less than or equal to the 'left', keys greater to the 'right'
This changes current key distribution in a tree so that all keys which are less
than or equal to current node's key go to the 'left' (lower index in
nodes's key array) and keys greater than current node's key go to the 'right'
(higher index in node's key array).

(Before, keys which are less got to the 'left' and keys which are
greater or equal go to the 'right'.)

This actually fixes bptree iterator to really start at defined key start
which can be just a key prefix - it doesn't need to match fully with the
key stored (e.g. all keys starting with KEY_OP_SYNC_C prefix that we use
as index to KV records which we need to sync with main kv store).