resource: ubridge: make it possible for cmd handler to span across more iterations of event loop
Make it possible for cmd handler to span across more iterations of event
loop to cover situations where we need to chunk the processing into
several steps, e.g. because of further information we need to collect
and hence we need to send out a message and wait for an answer using
existing event loop.