resource: ubridge: send full internal msg even for MSG_CATEGORY_CLIENT category
Avoid a special case for MSG_CATEGORY_CLIENT messages. Even though we are
sending a dummy internal message here with only the category field set
(the actualy message - the client request - is read out of the client connetion),
it's better to keep the messaging clean and do it the way we do for all the other
msg categories.