resource: ubridge: wait for scan cmd expbuf sync ack before sending resbuf
At the end of the execution of 'scan' command in worker process, we need to
sync content placed in cmd expbuf with main process' KV store. The worker
process must wait for main process to actually process the sync request
and wait for sync ack. When sync ack is received in worker process, it
can send cmd resbuf (that means exporting KV pairs to udev as result
of the scan command).

This avoids a race where executing two (or more) scan commands in
sequence could cause the scan command to see old main KV store state
where KV data from previous scan are not synchronized yet.