resource: ubridge: move protocol version check from _cmd_handler to _init_command
The _cmd_handler can be scheduled more than once if there are more data
expected to be processed. Move the protocol version check from _cmd_handler
to _init_command so the check is done only once for a command. Also, do the
version check as early as possible in _init_command.