modules: ucmd: blkid: unreserve keys on module exit
Note:
There used to be a comment:

  "TODO: Do not unreserve KVs in worker if we have modules preloaded in main process"

This is not that necessary, because module is destroyed after the sync point
with main KV store. That means, whatever we do to the db after the sync point,
this is kept in the snapshot only, including "unreservation" of reserved values
if we destroy modules in worker. We shouldn't ever be destroying modules
before sync point!

Maybe this is an extra useless job - to unreserve keys if we're going to
discard the whole KV store snapshot anyway right after, but I haven't found
an easy and quick way of determining whether this step can be surely skipped
or not. We can revisit this and add a patch for this later as optimization.