resource: ubridge: add sid_resource_type_ubridge_common subresource type
An instance of sid_resource_type_ubridge_common represents common
ubridge subtree that is used in both main and worker processes.

The common resource encapsulates:

  - KV store (sid_resource_type_kv_store)

  - aggregate with modules (sid_resource_type_aggregate), containing:
    - block modules (sid_resource_type_module_registry)
    - type modules (sid_resource_type_module_registry)

The instance of sid_resource_type_ubridge_common then contains a
struct sid_ucmd_common_ctx instance as the common resource's data
so it can be retrieved with sid_resource_get_data as usual.

This change makes it easier to share common ubridge subtree in both main
and worker processes - when a new worker is created, we just separate the
common resource (with children) and assign make it a child of the worker
process.

Also, this fixes an issue where the struct sid_ucmd_common_ctx was
actually a part of ubridge as ubridge->common_ctx. The problem was
that the ubridge->common_ctx was destroyed together with destroying
the ubridge instance in _worker_init_fn. Then we always had to
recreate a new common_ctx inside _init_command as sid_ucmd_ctx->common.

The flaw in design here was that we already preloaded modules in
the main process with the original struct sid_ucmd_common_ctx, because
struct sid_ucmd_common_ctx is registered as callback arg for for
'init', 'reset' and 'destroy' module callbacks. However, in worker
processes, this common_ctx was already destroyed and trying to use
it would result in use-after-free issue.

Creating a separate 'common' resource with struct sid_ucmd_common_ctx
inside resolves this issue because the 'common' resource is properly
inherited in worker processes from the main process and so the common
structure is still accessible even in the workers.