internal: bptree: fix reference counting when removing entries from the tree
Fix reference counting when removing an item from the tree in situation
where _redistribute_nodes is called for a node inside the tree during
node's key count underflow.

The _redistribute_nodes takes a key from neighbor node (also removing
it from the neighbor node), attaches it to the the current node and sets
the common parent key to the neighbor's key. In this case,
the _redistribute_nodes mistakenly called extra _ref_bkey which caused
the key reference counter to be off by one (+1).

Schematically and a bit simplified, this action can be drawn as:

          B                 A
        /  \    ----->       \
       A    C               B   C
     /  \  / \             /  \ / \
   .............         ............

Here, the "A" key had incorrect ref_count with extra +1 ref. The reason
for this bug is that _redistribute_nodes makes a simplification here and
it doesn't call unref for the "A" key in the neighbor node, but it just
cuts the number of keys by calling "neighbor->num_keys--" at the end of
_redistribute_nodes.

This is OK as far as we're inside _redistribute_nodes fn, so we just move
keys around there, keeping the overall ref_count same - we already do
this in other functions as well. Alternatively, we could be super
consistent and we could always call _ref_bkey, but then we'd also need
to call _unref_bkey in the neighbor node (...but then, for the consistency,
also changing all the other code that does similar key movements inside
the tree in a single function). But let's keep it simpler.

This patch also removes superfluous temporary variables that were
used in _redistribute_nodes and _coalesce_nodes just to type cast
generic pointer to a node pointer - we can do that directly and inline,
so writing:

  ((bptree_node_t *) neighbor->pointers[i])->parent = neighbor;

instead of:

  tmp = (bptree_node_t *) neighbor->pointers[i];
  tmp->parent = neighbor;