resource: ubridge: remove 'LYR' domain
Before, we recorded layering information based on sysfs (following the
"slaves" references for each device) in a separate LYR domain in KV
store. The LYR domain is now removed and the layering is recorded together
with each device's general "GMB" (group members) and "GIN" (group in)
records where all the other device references are stored.

This is for convenience so we can get all references for a single
device in one go (single access to the KV store).