resource: ubridge: assign new device ID as soon as the device is referenced
While we are checking sysfs for related devices ("slaves") and storing
this info inside SID db, use device ID for referencing all those related
devices. If any of them don't have a device ID assigned yet, then generate
and assign a new one right away. If we receive an event for such a device
later on, retrieve the already generated and assigned device ID for it
instead of creating a new one.