resource: ubridge: also check KV permissions in _kv_cb_delta_step
The _kv_cb_delta_step is a part of _kv_delta_set which is used to create
delta updates together with absolute diff calculations which are then
stored inside KV store. If we're managing records from outside the SID
core code (e.g. as part of sid_ucmd_group_* API calls), then we're
obliged to check the permissions.