base: buffer: close fd on failed vector buffer create
Vector buffers only store the iovecs in their memory. The buffers that
the iovecs point to still need to be malloced. If _buffer_vector_create
fails to malloc the memory to store the SID_BUFFER_MODE_SIZE_PREFIX
prefix buffer, it should close the fd. Otherwise it will leak when the
create fails.