base: buffer: make prefix vector buffer reset size consistent
When SID_BUFFER_MODE_SIZE_PREFIX buffers of both types are created,
their initial size is incremented to hold the prefix. When linear type
buffers are reset, their initial size calculation works the same way.
However, when vector type buffers are reset, the initial size is only
incremented if it was 0. Make this consistent with their creation and
the linear reset calculations.