buffer: Disallow releasing past used memory in release_mem
We need to check that the position we want to rewind to in actually
within the used memory. Otherwise, we can expose allowed but not used
garbage memory, or we can push usage.used past usage.allocated, which
can cause buffer errors when doing things like reading from a fd.