resource: ubridge: add separate print buffer in sid_ucmd_ctx
Before, we used generic buffer (sid_ucmd_ctx.common->gen_buf) for
printing results in various command handlers (_cmd_exec_version,
_cmd_exec_dbstats, _cmd_exec_resources, _cmd_exec_devices).

However, the generic buffer is supposed to be used only for
short-term temporary allocations, ideally not reaching out of
the scope of current function. Otherwise, we may get into an issue
where a pointer from the buffer is used and mark inside a buffer
then prevents taking another pointer from the buffer unless the
previous pointer is unbound (sid_buffer_unbind) or buffer memory
is rewound (sid_buffer_rewind).

This exact case caused an issue with
_cmd_exec_{version, dbstats, resources, devices} which used the
generic buffer for printing results, then taking the pointer to the
generic buffer memory with these results and adding that pointer to
the overall result buffer.

Later on, another part of the code tried to allocate memory from the
generic buffer and take the pointer (the sid_ucmd_mod_unreserve_kv
which needed this for composing and using the KV key temporarily).
It failed to do so because we were already using the pointer for the
buffer (the buffer inside was already marked by taking the previous
pointer for print results).

To avoid these issues, add a separate print buffer
(sid_ucmd_ctx.prn_buf) which is used for printing command results.
This also makes it cleaner so that the gen_buf stays there only for
the short-term temporary allocations as it was originally supposed to
and the separate prn_buf clearly holds the command printouts.